Extract plateau helper
