+++ src/game/index.js
@@ -1,4 +1,5 @@
 import * as THREE from 'three';
+import { createPlateau } from '../utils/map-utils.js';
 import { setupScene } from './setupScene.js';
 import { initSpawner, spawnUnit, spawnBuilding, setPathfinder as setSpawnerPathfinder } from './spawn.js';
 import { initPlacement, getPlacementMode, setPlacementMode, cancelPlacementMode, updateGhostBuilding, attemptPlacement, setPathfinder as setPlacementPathfinder } from './placement.js';
@@ -55,73 +56,19 @@ function openMapChunk() {
     const borderSize = 10;
     const plateauHeight = 2;
 
-    function createRampGeometry(width, length, height) {
-        const geometry = new THREE.BufferGeometry();
-        const vertices = new Float32Array([
-            -width / 2, height, 0,
-             width / 2, height, 0,
-            -width / 2, 0, -length,
-             width / 2, 0, -length,
-        ]);
-        geometry.setIndex([0, 2, 1, 2, 3, 1]);
-        geometry.setAttribute('position', new THREE.Float32BufferAttribute(vertices, 3));
-        geometry.computeVertexNormals();
-        return geometry;
-    }
-
     function addBorderPlateau(x, z, sizeX, sizeZ, orientation, withRamp = false) {
-        const plateauGeom = new THREE.BoxGeometry(sizeX, plateauHeight, sizeZ);
-        const plateau = new THREE.Mesh(plateauGeom, material);
-        plateau.position.set(x, plateauHeight / 2, z);
-        plateau.castShadow = true;
-        plateau.receiveShadow = true;
-        scene.add(plateau);
-
-        if (withRamp) {
-            const rampLength = 6;
-            const rampGeom = createRampGeometry(
-                orientation === 'north' || orientation === 'south' ? sizeX : sizeZ,
-                rampLength,
-                plateauHeight
-            );
-            const ramp = new THREE.Mesh(rampGeom, material);
-            ramp.name = 'ground';
-            switch (orientation) {
-                case 'north':
-                    ramp.rotation.y = 0;
-                    ramp.position.set(x, 0, z - sizeZ / 2 - rampLength / 2);
-                    break;
-                case 'south':
-                    ramp.rotation.y = Math.PI;
-                    ramp.position.set(x, 0, z + sizeZ / 2 + rampLength / 2);
-                    break;
-                case 'east':
-                    ramp.rotation.y = -Math.PI / 2;
-                    ramp.position.set(x + sizeX / 2 + rampLength / 2, 0, z);
-                    break;
-                case 'west':
-                    ramp.rotation.y = Math.PI / 2;
-                    ramp.position.set(x - sizeX / 2 - rampLength / 2, 0, z);
-                    break;
-                default:
-                    break;
-            }
-            ramp.castShadow = true;
-            ramp.receiveShadow = true;
-            scene.add(ramp);
-        }
-
-        const minX = x - sizeX / 2;
-        const maxX = x + sizeX / 2;
-        const minZ = z - sizeZ / 2;
-        const maxZ = z + sizeZ / 2;
-        const obstacle = {
-            collider: new THREE.Box3(
-                new THREE.Vector3(minX, 0, minZ),
-                new THREE.Vector3(maxX, plateauHeight, maxZ)
-            ),
-            getCollider() { return this.collider; }
-        };
+        const { meshes, collider } = createPlateau({
+            x,
+            z,
+            sizeX,
+            sizeZ,
+            height: plateauHeight,
+            orientation,
+            material,
+            withRamp
+        });
+        meshes.forEach(m => scene.add(m));
+        const obstacle = { collider, getCollider() { return this.collider; } };
         terrainObstacles.push(obstacle);
         collidableObjects.push(obstacle);
     }
